checkpoint. The algorithm looks to work correct now. I'm wondering if we can/should consolidate 'struct logrule' and 'struct draw_rule_format_' into one struct as they both appear to be output parameters for parse_ipt_logline.
The reason for wanting to do this is that libnet_log has a fairly limited method of passing
back and forth output/input to the callback function. One void *data gets set on nflog_callback_register,
by passing &logoutput to it we can write stuff to it from the callback but then we
cannot write stuff to rulefmt. We could ofcourse introduce a new struct consisting of
a logrule * and a rulefmt * but that gets kinda tacky.