vuurmuur-conf: redo service portrange mods
Portrange adds/mods/removals would lead to a bug in the edit service
logic, where prior changes weren't recorded. This was due to the way
the edit logic was restarted and intermediate results weren't properly
stored.

The new approach uses a new field for displaying current portranges
that can be refreshed. This way the restart logic is no longer needed
and thus removed.